Add the --delete option to aws s3 sync
This will remove files from the destination which are not present in the source.

Setting this allows "cf unbind-service" to remove an app from Prometheus targets
config. When files are removed from the active configuration in S3 we need that
change to the reflected to this Prometheus target configuration dir.